Create Touch-Averse-Robot-Arm
This project requires an OWI Arm and an hmc5883l mag sensor. The sensor detects and displays the magnetic field strengths in the x, y, and z directions and the magnitude of the field. The robot arm moves away from the magnetic field depending on which x y z direction has the strongest field strength.